+++ src/Nuclectic.Support.Helpers.InteropServices/Marshal/LateBoundMarshalSizeOf.cs
@@ -5,16 +5,16 @@ namespace Nuclectic.Support.Helpers.InteropServices.Marshal
 	public class LateBoundMarshalSizeOf
 		: IMarshalSizeOf
 	{
-		private readonly Func<Type, IntPtr> _sizeOf;
+		private readonly Func<Type, int> _sizeOf;
 
-		public LateBoundMarshalSizeOf(Func<Type, IntPtr> sizeOf)
+		public LateBoundMarshalSizeOf(Func<Type, int> sizeOf)
 		{
 			if (sizeOf == null) throw new ArgumentNullException("sizeOf");
 
 			_sizeOf = sizeOf;
 		}
 
-		public IntPtr? SizeOf(Type t)
+		public int? SizeOf(Type t)
 		{
 			return _sizeOf(t);
 		}
